hope last before release

+++ handlers/game_application/handle_form.py
@@ -6,6 +6,7 @@ from telebot.types import CallbackQuery, Message
 
 from controllers.game import GameController
 from controllers.game_application import GameApplicationController
+from handlers.game_application.states import GameApplicationStates
 from handlers.game_application.accept_form import generate_accept_form_markup
 from models import User
 from utils.message_helpers import get_user_text
@@ -46,9 +47,10 @@ async def handle_application_letter_no_data(
         await bot.edit_message_reply_markup(
             call.message.chat.id, call.message.message_id, reply_markup=None
         )
-        async with state.data() as data:
-            await send_application(bot, user, session, data["game_id"])
-        await state.delete()
+        if await state.get() == GameApplicationStates.letter.name:
+            async with state.data() as data:
+                await send_application(bot, user, session, data["game_id"])
+            await state.delete()
     except ApiTelegramException:
         pass
 
@@ -64,12 +66,13 @@ async def handle_application_cancel(
         await bot.edit_message_reply_markup(
             call.message.chat.id, call.message.message_id, reply_markup=None
         )
-        await state.delete()
-        await bot.send_message(
-            call.message.chat.id,
-            "Ты отменил заявку на игру. Давай поищем другие игры в канале - "
-            "https://t.me/SneakyDiceGames",
-        )
+        if await state.get() == GameApplicationStates.letter.name:
+            await state.delete()
+            await bot.send_message(
+                call.message.chat.id,
+                "Ты отменил заявку на игру. Давай поищем другие игры в канале - "
+                "https://t.me/SneakyDiceGames",
+            )
     except ApiTelegramException:
         pass
